Address PR review comments
- Rename DSA test registration files to DSATestRegistration.<Provider>.cs
- Remove dead #if NET guards around DSASignVerify_Span_* leaves (SSC tests
  only target NetCoreAppCurrent)
- Remove redundant parens and add blank lines per style nits in
  DSAKeyGeneration.cs and DSACreateTests.cs
- Move DSACngTests.cs from S.S.C.Cng/tests to S.S.C/tests so it can use
  DSACngProvider.Instance.SupportsFips186_3 directly; inline the three
  CNG-specific key blobs and remove them from TestData.cs
