+++ src/libraries/System.Security.Cryptography/tests/DSACngTests.Windows.cs
@@ -1,0 +1,119 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+using Test.Cryptography;
+using Xunit;
+
+namespace System.Security.Cryptography.Dsa.Tests
+{
+    public static class DsaCngTests
+    {
+        [Fact]
+        public static void TestImportV1Key()
+        {
+            using (CngKey key = CngKey.Import(s_key_DSA1024Key, CngKeyBlobFormat.GenericPrivateBlob))
+            {
+                VerifyImportedKey(key);
+                Assert.Equal(1024, key.KeySize);
+            }
+        }
+
+        [Fact]
+        public static void TestImportV2Key()
+        {
+            if (!DSACngProvider.Instance.SupportsFips186_3)
+            {
+                return;
+            }
+
+            using (CngKey key = CngKey.Import(s_key_DSA2048Key, CngKeyBlobFormat.GenericPrivateBlob))
+            {
+                VerifyImportedKey(key);
+                Assert.Equal(2048, key.KeySize);
+            }
+        }
+
+        private static void VerifyImportedKey(CngKey key)
+        {
+            Assert.Equal(new CngAlgorithm("DSA"), key.Algorithm);
+            Assert.Equal(CngAlgorithmGroup.Dsa, key.AlgorithmGroup);
+            Assert.Equal(CngExportPolicies.None, key.ExportPolicy);
+            Assert.True(key.IsEphemeral);
+            Assert.False(key.IsMachineKey);
+            Assert.Null(key.KeyName);
+            Assert.Equal(CngKeyUsages.AllUsages, key.KeyUsage);
+            Assert.Equal(IntPtr.Zero, key.ParentWindowHandle);
+            Assert.Equal(CngProvider.MicrosoftSoftwareKeyStorageProvider, key.Provider);
+
+            CngUIPolicy policy = key.UIPolicy;
+            Assert.Null(policy.CreationTitle);
+            Assert.Null(policy.Description);
+            Assert.Null(policy.FriendlyName);
+            Assert.Null(policy.UseContext);
+            Assert.Equal(CngUIProtectionLevels.None, policy.ProtectionLevel);
+            Assert.Null(key.UniqueName);
+        }
+
+        [Fact]
+        public static void TestImportExportRoundTrip()
+        {
+            using (CngKey key = CngKey.Import(s_key_DSA1024PublicKey, CngKeyBlobFormat.GenericPublicBlob))
+            {
+                byte[] reExported = key.Export(CngKeyBlobFormat.GenericPublicBlob);
+                Assert.Equal<byte>(s_key_DSA1024PublicKey, reExported);
+            }
+        }
+
+        [Fact]
+        public static void VerifyDefaultKeySize_Cng()
+        {
+            using (DSA dsa = new DSACng())
+            {
+                Assert.Equal(DSACngProvider.Instance.SupportsFips186_3 ? 2048 : 1024, dsa.KeySize);
+            }
+        }
+
+        private static readonly byte[] s_key_DSA1024Key = (
+            "4453505680000000000000AEE5A1D69A3F46E5225E5AFF69C5065D6AB1D6E2D9AB9515365DB8CAF4181C26C97F65BFB1931F" +
+            "EE7D865A61C79F41BE6ED77DF2E9BC7E1FF471FFDCFBEEA506234878C9871F5F85E650CDDE73F0B445BE5DD6A4A42A67F1F8" +
+            "9DB4C8CE799DD874AF06D8B8819210FF056A8DD76957C6B192AC590B13FC39000EF00AE2E77CD3AF869871BC758B8ABC2F97" +
+            "7FAADC9B3116C7F0A7B2970B621066464E6A5AA2AD528CB88DC0A2C5B8B92007C7C3A4C7472B47B0BB61F4F9A32A7F0F08B6" +
+            "8DF2C153B08AFD127C4483C6123C3375171C1040B8F0CFAAFDC58668AA819BB466DED1CF1A88026F8BFC73BA0D9B60218749" +
+            "88C90DBB2552B122D7A693F5183E7C455C7D984488D14DE99998DD14CC1B0420EBCC4E858762F2D61CEFDC4C46F72948AA24" +
+            "C7AF42CE544FE3E6446E408B6CFB1F81C212AE398BB6D2E69BD5B0AF35D2A8115A8609F3ADF101AE2288496CE05F9E49785D" +
+            "A01416787FBF270424874147995EC298854129E1469A265330EF4143DF02A20A8FB49C28D743EC9A06978BF7BA6AEE5D5FA4" +
+            "E177110CF26AF2791EBA2CEC63FB3B186A3E77ADC18F2070924C83BAED880DEC593AE1B9A6E41B4A33BC89A635F033C93B9C" +
+            "BA0FA7749E17").HexToByteArray();
+
+        private static readonly byte[] s_key_DSA1024PublicKey = (
+            "445350428000000000000003E9C7ECDB129947A0B6E1051311AF4410F3615518AE45325323742A6F08D3E69C9564C3909FC6" +
+            "A04DA3F643E048B96FF73CADAB870EA2F2A5E183691AEECC9694889E3BE02A81DCBBDB4DA810E38AB8072192AB26DDAD2F95" +
+            "4B5F0DF8DECFF1E72E56017F50E476BEE3702D89FAACEDB3FF6457C5545FA1D838299EDCF432785188313EF361322BF78AD3" +
+            "B491606BD6EBCA227571F2E14EFA9379AD9A78329C89FBB56835F6728513614A0037D042D82C486878A153C71F8C4890A9C9" +
+            "8CAB84FAF7D97C30081A2BE599CDD59E42BEC299BEFAFCBA4369653C81FF64F75976D8258E6CBD45A28C5774F90C29ED46BD" +
+            "178FEB031A003BE6870F69BD23BC1CC2A3DEA77506C7A96F2E2A4EAD74D31DB16EB696A8565577B95C2A5BC89058114BDDD1" +
+            "4D2AA790829B74BE2081080CFE4994EB1CBEE7F4B5001A8BC700EEE1511F25CDE5A08853F2DB10A477EFFB40E22B9B846DED" +
+            "A126CF159281A4889C0A59EF26A475ED193B853858944A7CA2750040DD8578418A38C733B7EFC6C18E7E9ABCCB5E0540D3A1" +
+            "2741968D89F7EFECBE709C7A451AACF070F6143ACC51B19B28CA49492C3A5EC322C9FFFA").HexToByteArray();
+
+        private static readonly byte[] s_key_DSA2048Key = (
+            "445056320001000001000000010000002000000020000000000001272B26541C50E35AFA386C959ED174D2DD70849A07D8EC" +
+            "E24838315A48E1679AD08125795599A6117EC55C2248303207C7C008CE1A86E9145D47810E621145871DF5C780BF53EAA45D" +
+            "8ADA28AACC40E101C4DB01645B4CA46A83754ECA82B94F48A04469CA40F61AB8E2CB10992A4062D5156C6F4E7FA50C6A3403" +
+            "3D6ED75BCFD68DCA234417B586FE1FB5B43751F6280C1D37AF6E07EB3F19A58E6762FDD73406CFA4B206A314D884A021F3F3" +
+            "167BDDA066F65C08E0E9CB37315DC428F5F14390F566BEE35E03D118210E2EF39272EA2EE1B060CAAA1D1919BB1DEE4149CE" +
+            "C45B5A4A995B984B0964A631DC44EABD918C75007C6844F03F3623BD27CAFE7B92F7AF041C834A58863482546972FBCB0F5A" +
+            "42625706D7088EE327BA8117AC115E671A8BCAD940D7A2E5A12021CC6E5133C9388120E58061C11C161078437CB15DBD1AE5" +
+            "4CF46DC427A79A34246EA8A8037038537119DA906808CFE18D827090C78A5D0061944465A955B764297DD0F0AC68E859414A" +
+            "08447CD91562A4B378DBDCCDC7B5118F3B5BC565D7F0198C24AD7CF6DEF4F3FF91302627BA6330FE7007BDAFEC34D644E010" +
+            "AA4F380F6322303430550E7D457041309C2F3E7323CE2EB8DE6DE31B68E07565BADD08A97159858725E474A9B2DC76556DA1" +
+            "1E124903BF21520CF288545ED4A53247E67EFB7620FE7F5DBEEF190569A88E29F50FAE343563FB49314FFEB7DFFA212AFF0E" +
+            "C280098A21AED0482057A97DA0AD4BA7E25A44376763EB035DCF065BDF187E2BF71E1A41BD1533F3F4D5432833BEB6485987" +
+            "2A28AB0A20C6E4C7F09843A443C1FD02444913B8362041E17962D8F29B77F4DF763ED97ACEE7978B7ECA8DE8C7D4EB7C46F4" +
+            "B274C3A780E4336A48E3526AF5F611A11FDB204069EA4212C8131D90D9158DDD29540B22B799A993652A8666E8BA2B6B9660" +
+            "02EAC030FE3D01CE4DA543ACAE89106564EA81FC1CEE358B613ED0D1A106BEE8458CCAB7D55C79532A80F8F5ED27C4FED459" +
+            "BC15B39737F3EDAF5F4F3260E34CFD75D790C3D404722A527B1D6AC52E704CD7FB4086820994CA88AA6DE32987882DFF6991" +
+            "CE0D641DE425FD0ACD614B824F75F58D84D96F8AA92E4C12F9233F0F70629295361993A7371E3352B6393072BB34F91C3B1D" +
+            "08C7B239ED66FDE828870C03AD24E3805D6E0B708AE0EF9E17474A1AEE754C6F865B943407B2CCD39B3E").HexToByteArray();
+    }
+}
