+++ src/libraries/System.Security.Cryptography/tests/DSACreateTests.cs
@@ -85,6 +85,7 @@ namespace System.Security.Cryptography.Tests
             {
                 return;
             }
+
             CreateWithParameters(DSATestData.GetDSA2048Params());
         }
